Move hints into get help panel
