+++ app/javascript/components/Editor.tsx
@@ -421,7 +421,6 @@ export default ({
               exerciseTitle={exercise.title}
             />
             <div className="options">
-              <Header.ActionHints assignment={panels.instructions.assignment} />
               <Header.ActionSettings
                 settings={settings}
                 setSettings={setSettings}
@@ -538,7 +537,10 @@ export default ({
                     mentorDiscussionsLink={links.mentorDiscussions}
                   />
                 ) : null}
-                <GetHelpPanel helpHtml={help.html}>
+                <GetHelpPanel
+                  assignment={panels.instructions.assignment}
+                  helpHtml={help.html}
+                >
                   {insider ? (
                     <ChatGPT.Wrapper
                       helpRecord={helpRecord}
