+++ app/javascript/components/editor/GetHelp/GetHelpPanelComponents/index.ts
@@ -1,0 +1,2 @@
+export { GetHelpPanelHelp as Help } from './GetHelpPanelHelp'
+export { GetHelpPanelHints as Hints } from './GetHelpPanelHints'
